requirements.txt
streamlit
openai
spacy
python-docx
PyPDF2
langdetect
reportlab
pandas
matplotlib
